added delete genre and author check so that if there are any books associated, it won't continue

+++ src/main/java/it/paa/service/AuthorService.java
@@ -75,8 +75,12 @@ public class AuthorService implements AuthorRepository {
     }
 
     @Override
-    public void delete(Long Id) throws NotFoundException {
+    public void delete(Long Id) throws NotFoundException, IllegalArgumentException {
         Author author = getById(Id);
+
+        if(!author.getBooks().isEmpty())
+            throw new IllegalArgumentException("cannot delete this author because it has associated books");
+
         entityManager.remove(author);
     }
 }
